refactor: Categories repository 추가

+++ src/main/kotlin/com/yorizori/yoremo/domain/categories/entity/Categories.kt
@@ -7,7 +7,7 @@ import jakarta.persistence.*
 class Categories(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val categoryId: Int? = null,
+    val categoryId: Long? = null,
 
     val name: String? = null,
